requirements.txt
# ==============================================================================
# DEPRESSION SCREENING SYSTEM - REQUIREMENTS
# ==============================================================================
# Python 3.12.7
# Install with: pip install -r requirements.txt
# ==============================================================================

# Core Deep Learning
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0

# Machine Learning
scikit-learn>=1.3.0
scipy>=1.10.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Web Interface
flask>=2.3.0
flask-cors>=4.0.0

# Video Processing (for webcam integration)
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
mediapipe>=0.10.0
Pillow>=10.0.0

# NLP
nltk>=3.8.0
spacy>=3.5.0

# Sentiment Analysis
textblob>=0.17.0
vaderSentiment>=3.3.2

# Progress Bars
tqdm>=4.65.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Logging
colorlog>=6.7.0

# Model Evaluation
scikit-plot>=0.3.7

# Utilities
pathlib>=1.0.1
pillow>=10.0.0

# Optional: For better performance
# numba>=0.57.0  # Uncomment if you want faster numpy operations

# ==============================================================================
# INSTALLATION NOTES:
# ==============================================================================
# 1. Create virtual environment (recommended):
#    python -m venv venv
#    venv\Scripts\activate (Windows)
#    source venv/bin/activate (Linux/Mac)
#
# 2. Install requirements:
#    pip install -r requirements.txt
#
# 3. Download spaCy model:
#    python -m spacy download en_core_web_sm
#
# 4. Download NLTK data:
#    python -c "import nltk; nltk.download('punkt'); nltk.download('stopwords')"
#
# 5. For OpenFace installation (see README.md for detailed instructions)
# ==============================================================================
